T2S con Mistral/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voxtral-terminal-backend"
version = "0.1.0"
description = "Backend terminal-first per download modello Voxtral, parsing markdown con datapizza-ai e voice cloning."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "datapizza-ai==0.1.0",
  "huggingface_hub>=1.8.0",
  "httpx>=0.28.1",
  "mistral-common>=1.10.0",
  "numpy>=1.26,<3",
  "python-dotenv>=1.2.2",
  "soundfile==0.13.1",
]

[project.optional-dependencies]
dev = ["pytest==9.0.2"]

[project.scripts]
voxtral-backend = "voxtral_terminal_backend.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
